backend/requirements.txt
# ── Core RAG & LLM ──────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
groq==0.9.0

# ── Vector Database (Chroma Cloud) ──────────────────────────────
chromadb==0.6.3

# ── Embeddings ──────────────────────────────────────────────────
sentence-transformers==3.0.1

# ── Stock Data ──────────────────────────────────────────────────
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4

# ── Backend API ─────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4

# ── Scheduler ───────────────────────────────────────────────────
schedule==1.2.2
apscheduler==3.10.4

# ── Utilities ────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2

# ── Testing ─────────────────────────────────────────────────────
pytest==8.3.2
pytest-cov==5.0.0
